refactor: split routes into a domain-organized routes/ package
routes.py mixed four concerns (site meta, scraping, billing, signup) in
one 238-line module. Split it into a routes/ package with one submodule
per concern:

- routes/meta.py    — index + health
- routes/scrape.py  — scrape, progress stream, download, cancel
- routes/billing.py — checkout, portal, usage status, webhook, key, signup
- routes/__init__.py aggregates them into the single router app.py mounts

Handlers moved verbatim; the duplicated tracker-id hashing in two scrape
handlers is now a shared _tracker_id_for() helper. The public import path
(link_content_scraper.routes.router) is unchanged. Tests retarget their
monkeypatches to the owning submodule (routes.scrape / routes.billing).
